// src/app/features/bills-invoices/purchase-invoices/purchase-invoices.types.ts
import { ButtonConfig } from '@components/button/button.types';
export interface IPurchaseInvoice {
  id: number;
  invoiceNumber: string;
  clientName: string;
  totalAmount: number;
  balance: number;
  status: string;
  date: string;
}
export interface IPageNavigationItem {
  key: string;
  buttonConfig: ButtonConfig;
}